+++ src/hotspot/share/gc/serial/genMemoryPools.cpp
@@ -24,11 +24,9 @@
 
 #include "precompiled.hpp"
 #include "gc/serial/generation.hpp"
-#include "gc/shared/genMemoryPools.hpp"
+#include "gc/serial/genMemoryPools.hpp"
 #include "gc/shared/space.hpp"
-#if INCLUDE_SERIALGC
 #include "gc/serial/defNewGeneration.hpp"
-#endif
 
 ContiguousSpacePool::ContiguousSpacePool(ContiguousSpace* space,
                                          const char* name,
@@ -49,8 +47,6 @@ MemoryUsage ContiguousSpacePool::get_memory_usage() {
 
   return MemoryUsage(initial_size(), used, committed, maxSize);
 }
-
-#if INCLUDE_SERIALGC
 
 SurvivorContiguousSpacePool::SurvivorContiguousSpacePool(DefNewGeneration* young_gen,
                                                          const char* name,
@@ -75,8 +71,6 @@ MemoryUsage SurvivorContiguousSpacePool::get_memory_usage() {
 
   return MemoryUsage(initial_size(), used, committed, maxSize);
 }
-
-#endif // INCLUDE_SERIALGC
 
 GenerationPool::GenerationPool(Generation* gen,
                                const char* name,
