+++ src/hotspot/share/gc/serial/serialHeap.cpp
@@ -33,6 +33,7 @@
 #include "gc/serial/cardTableRS.hpp"
 #include "gc/serial/defNewGeneration.inline.hpp"
 #include "gc/serial/genMarkSweep.hpp"
+#include "gc/serial/genMemoryPools.hpp"
 #include "gc/serial/markSweep.hpp"
 #include "gc/serial/serialHeap.hpp"
 #include "gc/serial/serialVMOperations.hpp"
@@ -50,7 +51,6 @@
 #include "gc/shared/gcTraceTime.inline.hpp"
 #include "gc/shared/gcVMOperations.hpp"
 #include "gc/shared/genArguments.hpp"
-#include "gc/shared/genMemoryPools.hpp"
 #include "gc/shared/locationPrinter.inline.hpp"
 #include "gc/shared/oopStorage.inline.hpp"
 #include "gc/shared/oopStorageParState.inline.hpp"
